On read error or EOF, consume buffered data
Previously when the read thread shutdown (which could be due to an error
or an EOF) we would set `read_error` to true. The next call to `read`
would then return `Eof, dropping buffered data.

This patch renames `read_error` to `read_thread_exit` which better
describes what it actually means and calls to `read` will return `Data`
if there is anything in the buffer.

Signed-off-by: David Scott <dave@recoil.org>